Suggest -libs suffix if rpm not found in rpm checker
For example, for sqlite we don't have the sqlite rpm installed, but
we do have sqlite-libs.

Actually there are many rpms we might have, see all the RPMs listed at
https://brewweb.engineering.redhat.com/brew/buildinfo?buildID=3758308

The comment suggests a way we could find them robustly, but for now
I don't think it's worth the effort.